refactor(bot): split index.ts by functionality
- Extract bot_ci merge blocking logic to mergeable.ts
- Keep index.ts clean with just app setup and handler registration
